Updated Decorator Pattern
Created a decorator pattern for ships and laser classes,  also using a factory pattern for instantiation.

Made zig-zagging possible for each ship.

Made it so that each time a ship is hit by the player, a random laser takes damage. When a laser takes full damage, it's destroyed. The player must destroy all lasers on a ship before the ship is destroyed.